还差锚点没修过

+++ server/clean.js
@@ -2,6 +2,8 @@ const fs = require('fs');
 const CONFIG = require('./config');
 const Utils = require('./utils')
 
-
+Utils.deleteFolder(CONFIG.CSS_PATH_DEV);
 Utils.deleteFolder('./public');
-fs.unlinkSync(CONFIG.CACHE_PATH);
+if( fs.existsSync(CONFIG.CACHE_PATH) ) {
+    fs.unlinkSync(CONFIG.CACHE_PATH);
+}
